pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a4design"
version = "0.1.0"
description = "UML Diagram Generation System using Multi-Agent Architecture"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"

keywords = ["uml", "diagrams", "ai", "multi-agent", "plantuml"]

dependencies = [
    "requests",
    "torch",
    "pydantic",
    "python-dotenv",
    "langchain-openai",
    "langchain-core",
    "langchain-community",
    "sentence-transformers",
    "langgraph",
    "python-Levenshtein",
    "tqdm",
    "scikit-learn",
    "numpy",
    "faiss-cpu",
]

[project.urls]
Repository = "https://github.com/yourusername/a4design.git"
Issues = "https://github.com/yourusername/a4design/issues"

[project.scripts]
a4design = "src.main:main"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.txt", "*.md"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]
known_third_party = ["langchain", "torch", "sklearn"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "faiss.*",
    "sentence_transformers.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]